para corregir busqueda propuestas

+++ src/main/java/edu/eci/cvds/persistence/UsuarioDAO.java
@@ -9,6 +9,7 @@ import edu.eci.cvds.entities.Rol;
 import edu.eci.cvds.entities.Usuario;
 import java.util.*;
 import java.lang.String;
+import java.math.BigInteger;
 
 public interface UsuarioDAO {
 	
@@ -20,6 +21,6 @@ public interface UsuarioDAO {
 	
 	public void actualizarUsuario(int rol, String correo) throws PersistenceException;
 	
-	public void agregarUsuario(int id, String tid, String nombre, int telefono, String correo, String clave, Rol rol, String estado) throws PersistenceException;
+	public void agregarUsuario(int id, String tid, String nombre, BigInteger  telefono, String correo, String clave, Rol rol, String estado) throws PersistenceException;
 
 }
